Move only: Move CDiskTxPos to its own file
